Add PressSource::is_secondary, is_tertiary
