+++ crates/kas-core/src/event/manager/press.rs
@@ -8,9 +8,76 @@
 #[allow(unused)]
 use super::{EventMgr, EventState, GrabMode, Response}; // for doc-links
 use super::{MouseGrab, Pending, TouchGrab};
-use crate::event::{CursorIcon, PressSource};
+use crate::event::{CursorIcon, MouseButton};
 use crate::geom::{Coord, Offset};
 use crate::{Action, WidgetId};
+
+/// Source of `EventChild::Press`
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub enum PressSource {
+    /// A mouse click
+    ///
+    /// Arguments: `button, repeats`.
+    ///
+    /// The `repeats` argument is used for double-clicks and similar. For a
+    /// single-click, `repeats == 1`; for a double-click it is 2, for a
+    /// triple-click it is 3, and so on (without upper limit).
+    ///
+    /// For `PressMove` and `PressEnd` events delivered with a mouse-grab,
+    /// both arguments are copied from the initiating `PressStart` event.
+    /// For a `PressMove` delivered without a grab (only possible with pop-ups)
+    /// a fake `button` value is used and `repeats == 0`.
+    Mouse(MouseButton, u32),
+    /// A touch event (with given `id`)
+    Touch(u64),
+}
+
+impl PressSource {
+    /// Returns true if this represents the left mouse button or a touch event
+    #[inline]
+    pub fn is_primary(self) -> bool {
+        match self {
+            PressSource::Mouse(button, _) => button == MouseButton::Left,
+            PressSource::Touch(_) => true,
+        }
+    }
+
+    /// Returns true if this represents the right mouse button
+    #[inline]
+    pub fn is_secondary(self) -> bool {
+        match self {
+            PressSource::Mouse(MouseButton::Right, _) => true,
+            _ => false,
+        }
+    }
+
+    /// Returns true if this represents the middle mouse button
+    #[inline]
+    pub fn is_tertiary(self) -> bool {
+        match self {
+            PressSource::Mouse(MouseButton::Middle, _) => true,
+            _ => false,
+        }
+    }
+
+    /// Returns true if this represents a touch event
+    #[inline]
+    pub fn is_touch(self) -> bool {
+        matches!(self, PressSource::Touch(_))
+    }
+
+    /// The `repetitions` value
+    ///
+    /// This is 1 for a single-click and all touch events, 2 for a double-click,
+    /// 3 for a triple-click, etc. For `PressMove` without a grab this is 0.
+    #[inline]
+    pub fn repetitions(self) -> u32 {
+        match self {
+            PressSource::Mouse(_, repetitions) => repetitions,
+            PressSource::Touch(_) => 1,
+        }
+    }
+}
 
 /// Details of press events
 #[crate::autoimpl(Deref using self.source)]
